artifacts/code/experiment/requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
